Add files via upload
If your score is 57 or below, you are a low LPC, which suggests that you are task motivated. If your score is within the range of 58 to 63, you are a middle LPC, which means you are independent. People who score 64 or above are called high LPCs, and they are thought to be more relationship motivated.